requirements.txt
pandas
tabulate
click
tqdm
GitPython
